+++ docs/source/conf.py
@@ -63,6 +63,7 @@ intersphinx_mapping = {
     "pytorch_lightning": ("https://pytorch-lightning.readthedocs.io/en/latest", None),
     # Just leaving this here for when we do finally add a deepspeed integration:
     #  "deepspeed": ("https://deepspeed.readthedocs.io/en/latest/", None),
+    "flax": ("https://flax.readthedocs.io/en/latest", None),
     "fairscale": ("https://fairscale.readthedocs.io/en/latest/", None),
     "datasets": ("https://huggingface.co/docs/datasets/master/en", None),
     "transformers": ("https://huggingface.co/docs/transformers/master/en", None),