+++ tango/integrations/flax/__init__.py
@@ -1,0 +1,32 @@
+from tango.common.exceptions import IntegrationMissingError
+
+try:
+    import flax
+except ModuleNotFoundError:
+    raise IntegrationMissingError("flax")
+
+__all__ = [
+    "DataLoader",
+    "FlaxDataLoader",
+    "LRScheduler",
+    "Model",
+    "Optimizer",
+    "FlaxTrainStep",
+    "FlaxFormat",
+    "TrainCallback",
+    "EvalCallback",
+    "FlaxWrapper",
+    "TrainConfig",
+    "FlaxEvalStep",
+]
+
+from .data import DataLoader, FlaxDataLoader
+from .eval import FlaxEvalStep
+from .eval_callback import EvalCallback
+from .format import FlaxFormat
+from .model import Model
+from .optim import LRScheduler, Optimizer
+from .train import FlaxTrainStep
+from .train_callback import TrainCallback
+from .train_config import TrainConfig
+from .wrapper import FlaxWrapper